requirements.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fsspec==2024.6.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.0
tornado==6.4
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.12.1
tzdata==2024.1
wcwidth==0.2.13
